ref: create navbar component
